feat: VitestのIn-source Testingに移行
- __tests__/ディレクトリのテストコードをsrc/配下の各ファイルに移動
- vitest.config.tsにincludeSource設定を追加してIn-source Testing有効化
- src/utils/test-utils.tsxを作成してChakra UIのテストヘルパーを提供
- 全テストが正常に実行されることを確認済み
