+++ src/components/page/index/accounts.tsx
@@ -92,3 +92,67 @@ const Accounts = (props: Props) => {
 };
 
 export default Accounts;
+
+if (import.meta.vitest) {
+	const { describe, expect, it } = import.meta.vitest;
+	const { screen } = await import("@testing-library/react");
+	const { renderWithChakra } = await import("@/utils/test-utils");
+
+	describe("Accounts component", () => {
+		it("renders correctly", () => {
+			const mockProps = {
+				twitter: "https://twitter.com/test_twitter",
+				github: "https://github.com/test_github",
+				qiita: "https://qiita.com/test_qiita",
+				speakerDeck: "https://speakerdeck.com/test_speakerdeck",
+				note: "https://note.com/test_note",
+				hatenaBlog: "https://hatenablog.com/test_hatenaBlog",
+				zenn: "https://zenn.dev/test_zenn",
+			};
+
+			const accountsMap: SNSAccount[] = [
+				{
+					name: "Twitter",
+					url: mockProps.twitter,
+				},
+				{
+					name: "GitHub",
+					url: mockProps.github,
+				},
+				{
+					name: "Zenn",
+					url: mockProps.zenn,
+				},
+				{
+					name: "Qiita",
+					url: mockProps.qiita,
+				},
+				{
+					name: "SpeakerDeck",
+					url: mockProps.speakerDeck,
+				},
+			];
+
+			const blogsMap: BlogAccount[] = [
+				{
+					name: "HatenaBlog",
+					url: mockProps.hatenaBlog,
+				},
+				{
+					name: "note",
+					url: mockProps.note,
+				},
+			];
+
+			renderWithChakra(<Accounts {...mockProps} />);
+
+			for (const { name, url } of accountsMap) {
+				expect(screen.getByRole("link", { name })).toHaveAttribute("href", url);
+			}
+
+			for (const { name, url } of blogsMap) {
+				expect(screen.getByRole("link", { name })).toHaveAttribute("href", url);
+			}
+		});
+	});
+}
