+++ src/components/page/index/blog.tsx
@@ -21,3 +21,25 @@ const Blog = ({ name, url }: BlogProps) => {
 };
 
 export default Blog;
+
+if (import.meta.vitest) {
+	const { describe, expect, it } = import.meta.vitest;
+	const { screen } = await import("@testing-library/react");
+	const { renderWithChakra } = await import("@/utils/test-utils");
+
+	describe("Blog component", () => {
+		it("renders correctly", () => {
+			const mockProps = {
+				name: "HatenaBlog",
+				url: "http://teitei-tk.hatenablog.com/",
+			};
+
+			renderWithChakra(<Blog {...mockProps} />);
+
+			expect(screen.getByRole("link", { name: "HatenaBlog" })).toHaveAttribute(
+				"href",
+				"http://teitei-tk.hatenablog.com/",
+			);
+		});
+	});
+}
