+++ src/components/page/index/profile.tsx
@@ -71,3 +71,40 @@ const Profile = (props: Props) => {
 };
 
 export default Profile;
+
+if (import.meta.vitest) {
+	const { describe, expect, it } = import.meta.vitest;
+	const { screen } = await import("@testing-library/react");
+	const { renderWithChakra } = await import("@/utils/test-utils");
+
+	describe("Profile component", () => {
+		it("renders correctly", () => {
+			const mockProps = {
+				name: "Test Name",
+				bio: "Test Bio",
+				email: "test@example.com",
+				avatarURL: "http://example.com/avatar.jpg",
+				twitter: "https://twitter.com/test_twitter",
+			};
+
+			renderWithChakra(<Profile {...mockProps} />);
+
+			expect(screen.getByText("name")).toBeInTheDocument();
+			expect(screen.getByText(`@${mockProps.name}`)).toBeInTheDocument();
+			expect(screen.getByText("biography")).toBeInTheDocument();
+			expect(screen.getByText(mockProps.bio)).toBeInTheDocument();
+			expect(screen.getByText("Contact")).toBeInTheDocument();
+			expect(screen.getByText(mockProps.email)).toBeInTheDocument();
+			expect(screen.getByRole("link", { name: mockProps.email })).toHaveAttribute(
+				"href",
+				"https://twitter.com/test_twitter",
+			);
+
+			expect(
+				screen.getByAltText(
+					"teitei-tkとして活動しているアカウントのアバター画像。タバコをくわえた茶髪の少女の横顔イラスト",
+				),
+			).toHaveAttribute("src", "http://example.com/avatar.jpg");
+		});
+	});
+}
