+++ src/pages/index.tsx
@@ -112,3 +112,96 @@ export const getStaticProps: GetStaticProps = async (_) => {
 };
 
 export default IndexPage;
+
+if (import.meta.vitest) {
+	const { describe, expect, it } = import.meta.vitest;
+	const { screen } = await import("@testing-library/react");
+	const { renderWithChakra } = await import("@/utils/test-utils");
+
+	describe("IndexPage component", () => {
+		it("renders correctly", () => {
+			const mockProps: IndexPageProps = {
+				siteMetadata: {
+					name: "teitei-tk.com",
+					repositoryURL: "https://github.com/teitei-tk/teitei-tk.com",
+					user: {
+						name: "teitei-tk",
+						bio: "Software Engineer writing Ruby, Python, PHP, JavaScript, Go",
+						avatarURL: "https://avatars3.githubusercontent.com/u/1324680?v=4",
+						email: "teitei.tk@gmail.com",
+					},
+					accounts: {
+						twitter: "https://twitter.com/teitei_tk",
+						github: "https://github.com/teitei-tk",
+						qiita: "https://qiita.com/teitei_tk",
+						speakerDeck: "https://speakerdeck.com/teitei",
+						note: "https://note.com/teitei_tk",
+						hatenaBlog: "http://teitei-tk.hatenablog.com/",
+						zenn: "https://zenn.dev/teitei_tk",
+					},
+				},
+			};
+
+			renderWithChakra(<IndexPage {...mockProps} />);
+
+			// Check if the user name is rendered correctly
+			expect(screen.getByText(mockProps.siteMetadata.name)).toBeInTheDocument();
+
+			const accountsMap = [
+				{
+					name: "Twitter",
+					url: mockProps.siteMetadata.accounts.twitter,
+				},
+				{
+					name: "GitHub",
+					url: mockProps.siteMetadata.accounts.github,
+				},
+				{
+					name: "Zenn",
+					url: mockProps.siteMetadata.accounts.zenn,
+				},
+				{
+					name: "Qiita",
+					url: mockProps.siteMetadata.accounts.qiita,
+				},
+				{
+					name: "SpeakerDeck",
+					url: mockProps.siteMetadata.accounts.speakerDeck,
+				},
+			];
+
+			const blogsMap = [
+				{
+					name: "HatenaBlog",
+					url: mockProps.siteMetadata.accounts.hatenaBlog,
+				},
+				{
+					name: "note",
+					url: mockProps.siteMetadata.accounts.note,
+				},
+			];
+
+			for (const { name, url } of accountsMap) {
+				if (name === "GitHub") {
+					const links = screen.getAllByText(/GitHub/);
+					expect(links.length).toBe(2);
+					expect(links[0]).toHaveAttribute(
+						"href",
+						mockProps.siteMetadata.accounts.github,
+					);
+					expect(links[1]).toHaveAttribute(
+						"href",
+						mockProps.siteMetadata.repositoryURL,
+					);
+					return;
+				}
+
+				expect(screen.getByRole("link", { name })).toHaveAttribute("href", url);
+			}
+
+			for (const { name, url } of blogsMap) {
+				expect(screen.getByRole("link", { name })).toHaveAttribute("href", url);
+			}
+		});
+	});
+}
