+++ src/utils/test-utils.tsx
@@ -1,3 +1,6 @@
+/**
+ * テスト用のユーティリティ関数
+ */
 import { ChakraProvider, defaultSystem } from "@chakra-ui/react";
 import { render } from "@testing-library/react";
 import type React from "react";
